agentic-sre-demo/requirements.txt
# Core framework
agno>=0.1.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Data processing
pydantic>=2.5.0
pandas>=2.1.0
numpy>=1.24.0

# Observability and monitoring
prometheus-client>=0.19.0
elasticsearch>=8.11.0
opensearch-py>=2.4.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Async support
asyncio-mqtt>=0.16.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.7

# Time handling
python-dateutil>=2.8.2
pytz>=2023.3

# Configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Security
cryptography>=41.0.0
passlib>=1.7.4

# Additional utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Dashboard dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0
requests>=2.31.0 